requirements.txt
anyio==3.6.2
base58==2.1.1
certifi==2022.9.24
Deprecated==1.2.13
ECPy==1.2.5
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.4
rfc3986==1.5.0
sniffio==1.3.0
types-Deprecated==1.2.9
typing_extensions==4.4.0
websockets==10.4
wrapt==1.14.1
xrpl-py==1.7.0
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.0.1
h11==0.14.0
idna==3.4
numpy==1.23.4
openai==0.25.0
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.1
pandas-stubs==1.5.1.221024
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.12
wsproto==1.2.0